+++ sdk/eventhub/event-hubs/src/eventHubReceiver.ts
@@ -504,43 +504,14 @@ export class EventHubReceiver extends LinkEntity {
    * @returns
    */
   async close(abortSignal?: AbortSignalLike): Promise<void> {
-    return new Promise<void>(async (resolve, reject) => {
-      if (this._receiver) {
-        const rejectOnAbort = () => {
-          const desc: string =
-            `[${this._context.connectionId}] The close request on the receiver` +
-            `to "${this.address}" has been cancelled by the user.`;
-          log.error(desc);
-          const error = new AbortError(`The close operation has been cancelled by the user.`);
-          reject(error);
-        };
-
-        const onAbort = () => {
-          abortSignal!.removeEventListener("abort", onAbort);
-          rejectOnAbort();
-        };
-
-        if (abortSignal) {
-          // the aborter may have been triggered between request attempts
-          // so check if it was triggered and reject if needed.
-          if (abortSignal.aborted) {
-            return rejectOnAbort();
-          }
-          abortSignal.addEventListener("abort", onAbort);
-        }
-        log.sender(
-          "[%s] Closing the Receiver for the entity '%s'.",
-          this._context.connectionId,
-          this._context.config.entityPath
-        );
-        const receiverLink = this._receiver;
-        this._deleteFromCache();
-        await this._closeLink(receiverLink);
-        return resolve();
-      } else {
-        return resolve();
-      }
-    });
+    log.receiver(
+      "[%s] Closing the Receiver for the entity '%s'.",
+      this._context.connectionId,
+      this._context.config.entityPath
+    );
+    const receiverLink = this._receiver;
+    this._deleteFromCache();
+    await this._closeLink(receiverLink, abortSignal);
   }
 
   /**
